:sparkles: RunningData 모듈 분리
