+++ presentation/build.gradle
@@ -51,6 +51,7 @@ android {
 dependencies {
 
     implementation project(":domain")
+    implementation project(":runningdata")
 
     implementation "androidx.core:core-ktx:$coreKtxVersion"
     implementation "androidx.appcompat:appcompat:$appcompatVersion"
