+++ presentation/src/main/java/com/whyranoid/presentation/running/Event.kt
@@ -1,6 +1,6 @@
 package com.whyranoid.presentation.running
 
-import com.whyranoid.presentation.running.runningdatamanager.RunningFinishData
+import com.whyranoid.runningdata.model.RunningFinishData
 
 sealed interface Event {
     data class FinishButtonClick(val runningFinishData: RunningFinishData) : Event
