+++ presentation/src/main/java/com/whyranoid/presentation/runningstart/RunningStartFragment.kt
@@ -18,11 +18,11 @@ import com.whyranoid.presentation.base.BaseFragment
 import com.whyranoid.presentation.databinding.FragmentRunningStartBinding
 import com.whyranoid.presentation.running.RunningActivity
 import com.whyranoid.presentation.running.RunningViewModel.Companion.RUNNING_FINISH_DATA_KEY
-import com.whyranoid.presentation.running.runningdatamanager.RunningFinishData
-import com.whyranoid.presentation.running.runningdatamanager.RunningState
 import com.whyranoid.presentation.util.getSerializableData
 import com.whyranoid.presentation.util.gpsstate.GPSState
 import com.whyranoid.presentation.util.repeatWhenUiStarted
+import com.whyranoid.runningdata.model.RunningFinishData
+import com.whyranoid.runningdata.model.RunningState
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
