+++ presentation/src/main/java/com/whyranoid/presentation/runningstart/RunningStartViewModel.kt
@@ -3,8 +3,8 @@ package com.whyranoid.presentation.runningstart
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.whyranoid.domain.usecase.GetRunnerCountUseCase
-import com.whyranoid.presentation.running.runningdatamanager.RunningDataManager
 import com.whyranoid.presentation.util.networkconnection.NetworkConnectionStateHolder
+import com.whyranoid.runningdata.RunningDataManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
