+++ presentation/src/main/res/layout/activity_running.xml
@@ -5,7 +5,7 @@
 
     <data>
 
-        <import type="com.whyranoid.presentation.running.runningdatamanager.RunningState" />
+        <import type="com.whyranoid.runningdata.model.RunningState" />
 
         <import type="android.view.View" />
