+++ runningdata/build.gradle
@@ -1,0 +1,41 @@
+plugins {
+    id 'com.android.library'
+    id 'org.jetbrains.kotlin.android'
+}
+
+android {
+    namespace 'com.whyranoid.runningdata'
+    compileSdk 33
+
+    defaultConfig {
+        minSdk 23
+        targetSdk 33
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = '1.8'
+    }
+}
+
+    dependencies {
+
+        implementation "androidx.core:core-ktx:$coreKtxVersion"
+        implementation "androidx.appcompat:appcompat:$appcompatVersion"
+        implementation "com.google.android.material:material:$materialVersion"
+        testImplementation "junit:junit:$junitVersion"
+        androidTestImplementation "androidx.test.ext:junit:$junitUiVersion"
+        androidTestImplementation "androidx.test.espresso:espresso-core:$espressoCoreVersion"
+    }
