+++ runningdata/src/main/java/com/whyranoid/runningdata/RunningDataManager.kt
@@ -1,9 +1,13 @@
-package com.whyranoid.presentation.running.runningdatamanager
+package com.whyranoid.runningdata
 
 import android.location.Location
+import com.whyranoid.runningdata.model.RunningData
+import com.whyranoid.runningdata.model.RunningFinishData
+import com.whyranoid.runningdata.model.RunningPosition
+import com.whyranoid.runningdata.model.RunningState
+import com.whyranoid.runningdata.model.toRunningFinishData
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import java.util.UUID
 
 class RunningDataManager {
     private var _runningState = MutableStateFlow<RunningState>(
@@ -114,50 +118,3 @@ class RunningDataManager {
         }
     }
 }
-
-// TODO : 모델 파일 분리
-data class RunningData(
-    val startTime: Long = 0L,
-    val runningTime: Int = 0,
-    val totalDistance: Double = 0.0,
-    val pace: Double = 0.0,
-    val runningPositionList: List<List<RunningPosition>> = listOf(emptyList()),
-    val lastLocation: Location? = null
-)
-
-data class RunningFinishData(
-    val runningHistory: RunningHistoryModel,
-    val runningPositionList: List<List<RunningPosition>>
-) : java.io.Serializable
-
-sealed interface RunningState {
-    val runningData: RunningData
-
-    data class NotRunning(override val runningData: RunningData = RunningData()) : RunningState
-
-    data class Running(override val runningData: RunningData) : RunningState
-
-    data class Paused(override val runningData: RunningData) : RunningState
-}
-
-fun RunningData.toRunningFinishData() =
-    RunningFinishData(
-        RunningHistoryModel(
-            historyId = UUID.randomUUID().toString(),
-            startedAt = startTime,
-            finishedAt = System.currentTimeMillis(),
-            totalRunningTime = runningTime,
-            pace = pace,
-            totalDistance = totalDistance
-        ),
-        runningPositionList
-    )
-
-data class RunningHistoryModel(
-    val historyId: String = "",
-    val startedAt: Long = 0L,
-    val finishedAt: Long = 0L,
-    val totalRunningTime: Int = 0,
-    val pace: Double = 0.0,
-    val totalDistance: Double = 0.0
-) : java.io.Serializable
