+++ runningdata/src/main/java/com/whyranoid/runningdata/model/RunningFinishData.kt
@@ -1,0 +1,6 @@
+package com.whyranoid.runningdata.model
+
+data class RunningFinishData(
+    val runningHistory: RunningHistoryModel,
+    val runningPositionList: List<List<RunningPosition>>
+) : java.io.Serializable
