+++ runningdata/src/main/java/com/whyranoid/runningdata/model/RunningPosition.kt
@@ -1,0 +1,6 @@
+package com.whyranoid.runningdata.model
+
+data class RunningPosition(
+    val latitude: Double,
+    val longitude: Double
+) : java.io.Serializable
