+++ runningdata/src/main/java/com/whyranoid/runningdata/model/RunningState.kt
@@ -1,0 +1,11 @@
+package com.whyranoid.runningdata.model
+
+sealed interface RunningState {
+    val runningData: RunningData
+
+    data class NotRunning(override val runningData: RunningData = RunningData()) : RunningState
+
+    data class Running(override val runningData: RunningData) : RunningState
+
+    data class Paused(override val runningData: RunningData) : RunningState
+}
