+++ settings.gradle
@@ -25,3 +25,4 @@ include ':domain'
 include ':data'
 include ':presentation'
 include ':signin'
+include ':runningdata'
